requirements.txt
Django
djangorestframework
django-cors-headers
djoser
drf-yasg
setuptools
djangorestframework-simplejwt
